requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.1
pydantic>=2.4.2
openai>=1.2.0
python-json-logger>=2.0.7
retry>=0.9.2 